+++ src/DomToArray.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace VaclavVanik\DomToArray;
 
 use DOMCdataSection;
-use DOMComment;
 use DOMDocument;
 use DOMElement;
 use DOMText;
@@ -138,10 +137,6 @@ class DomToArray
         $childNamesCount = $this->childNamesCount($element);
 
         foreach ($element->childNodes as $childNode) {
-            if ($childNode instanceof DOMComment) {
-                continue;
-            }
-
             if ($childNode instanceof DOMCdataSection) {
                 $result[self::KEY_VALUE] = ($result[self::KEY_VALUE] ?? '') . $childNode->data;
                 continue;
@@ -149,6 +144,10 @@ class DomToArray
 
             if ($childNode instanceof DOMText) {
                 $result[self::KEY_VALUE] = ($result[self::KEY_VALUE] ?? '') . $childNode->textContent;
+                continue;
+            }
+
+            if (! ($childNode instanceof DOMElement)) {
                 continue;
             }